requirements.txt
boto3==1.42.66
botocore==1.42.66
fastavro==1.12.1
jmespath==1.1.0
numpy==2.4.3
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
